// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/api/meals.service.ts
export interface GetMealsRequestParams {
    version: string;
    page?: number;
    limit?: number;
    dateFrom?: string;
    dateTo?: string;
    mealTypes?: string;
    caloriesFrom?: number;
    caloriesTo?: number;
    hasImage?: boolean;
    hasAiSession?: boolean;
}